requirements.txt
python-telegram-bot
python-dotenv


# Extra
pandas-datareader